+++ x/perp/keeper/v2/clearing_house.go
@@ -520,8 +520,6 @@ func (k Keeper) afterPositionUpdate(
 	}
 
 	if !positionResp.Position.Size_.IsZero() {
-		k.Positions.Insert(ctx, collections.Join(market.Pair, traderAddr), *positionResp.Position)
-
 		spotNotional, err := PositionNotionalSpot(amm, *positionResp.Position)
 		if err != nil {
 			return err
@@ -556,6 +554,10 @@ func (k Keeper) afterPositionUpdate(
 	transferredFee, err := k.transferFee(ctx, market.Pair, traderAddr, positionResp.ExchangedNotionalValue)
 	if err != nil {
 		return err
+	}
+
+	if !positionResp.Position.Size_.IsZero() {
+		k.Positions.Insert(ctx, collections.Join(market.Pair, traderAddr), *positionResp.Position)
 	}
 
 	// calculate positionNotional (it's different depends on long or short side)